Remove cycle
